Separate send::{v1,v2} modules
This reduces the number of feature gates and works toward a crate
where the features are additive and potentially both available
at the same time.

Right now they're still mutually exclusive (v1 is enabled by not(v2))
